toi uu he thong

+++ app/Http/Controllers/AttendanceController.php
@@ -105,8 +105,13 @@ class AttendanceController
         {
             return back();
         }
-        Attendance::findOrFail($id)->delete();
-        return back()->with('success', 'Xóa dữ liệu chấm công thành công');
+        $exists = DB::table('attendances')->where('id', $id)->exists();
+        if($exists)
+        {
+            Attendance::findOrFail($id)->delete();
+            return back()->with('success', 'Xóa dữ liệu chấm công thành công');
+        }
+        return back();
     }
 
     public function confirm($id)
@@ -115,7 +120,6 @@ class AttendanceController
         {
             return back();
         }
-        
         DB::table('attendances')->where('id', $id)->update(['confirm' => 'yes']);
         return redirect('/qlcl/attendances')->with('success', 'Xác nhận chấm công thành công');
     }
